requirements.txt
Django==4.1
psycopg2>=2.8
djangorestframework==3.13.1
PyJWT==2.4.0
django-cors-headers==3.13.0
drf_yasg
celery
redis
django-celery-beat
django-timezone-field==4.0
Pillow
pytest
pytest-django
